+++ tests/Set/Silverstripe60/Silverstripe60Test.php
@@ -1,0 +1,25 @@
+<?php
+declare(strict_types=1);
+namespace Netwerkstatt\SilverstripeRector\Tests\Set\Silverstripe60;
+
+use Rector\Testing\PHPUnit\AbstractRectorTestCase;
+use PHPUnit\Framework\Attributes\DataProvider;
+
+class Silverstripe60Test extends AbstractRectorTestCase
+{
+    #[DataProvider('provideData')]
+    public function test(string $filePath): void
+    {
+        $this->doTestFile($filePath);
+    }
+
+    public static function provideData(): \Iterator
+    {
+        return self::yieldFilesFromDirectory(__DIR__ . '/Fixture');
+    }
+
+    public function provideConfigFilePath(): string
+    {
+        return __DIR__ . '/config/configured_rule.php';
+    }
+}